requirements.txt
# 飞书企业级长程协作记忆系统 - 依赖
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
apscheduler>=3.10.0
httpx>=0.25.0
openai>=1.6.0
python-dateutil>=2.8.2
rich>=13.0.0
jinja2>=3.1.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
